chatbot/app/requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
langgraph
python3-dotenv